Output note/magnitude as a time series CSV
I can identify (most of) the proper notes in Canon in D, but there's a
lot of noisy data around the notes. (Adjacent notes also show up for
lower-pitch notes, and some don't meet the given threshold).

Might be able to produce a good heuristic with the following:
- Adjacent notes are unlikely to be played at the same time, so
  determine which notes are being played by looking at which notes are
  louder than their neighbours in a frame.
- Determine when a note "starts" by looking for peaks in the note's time
  series -- so, if it was quieter last frame, and quieter in the next
  frame, then it probably just got played.
- Compute a threshold based on note frequency, and delete anything under
  that threshold.

There might be some necessary work with identifying harmonics, but I'll
cross that bridge when I come to it.